Add automatic rebuild/reload when files change.
If the server is started with reload=True, any change in the files will
trigger a server reload which, in turn triggers a rebuild and updates
the server start time in the _state variable.  The client side monitors
the server start time and will automatically reload itself when it sees
a server start time newer than the previous.