Support HTML comment tag.
The comment tag requires a special closing, " -->" and doesn't take any
attributes. It's now supported and corresponding doctests were added. I
also changed the signature of the __init__ function: attrs and content
are now keyword arguments with defaults of {} and [], respectively. This
makes it just a little easier to create initially empty elements and add
to them in subsequent code.