Add two new demo files
minimal_allinone.py removes most of the content generation and update
logic from allinone.py. It simply presents a 'Hello' banner and updates
an onscreen counter with values supplied by the server.
serverless.py removes all the server and auto-reaload code. It simply
compiles the html and js and uses Python's webbrowser module to load the
file into the default broweser. The display is identical to
minimal_allinone but the counter is updated directly from a JS function
with no server callbacks.